requirements.txt
boto3>=1.38.36
pdfminer.six>=20250506
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0 